requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
anthropic==0.85.0
pyoxigraph==0.5.6
python-dotenv==1.2.2
